+++ src/main/java/com/yanny/ytech/configuration/renderer/CraftingWorkspaceRenderer.java
@@ -64,6 +64,7 @@ public class CraftingWorkspaceRenderer implements BlockEntityRenderer<CraftingWo
                             ItemStack itemStack = items.get(i);
 
                             if (position == null || itemStack.isEmpty()) {
+                                i++;
                                 continue;
                             }